ファイル内の関数をモジュールに分割
